Store関連の型定義を別ファイルに切り出し
